requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
kafka-python==2.0.2
mypy==1.13.0
mypy-extensions==1.0.0
py4j==0.10.9.5
pyspark==3.3.0
python-dotenv==1.0.1
requests==2.32.3
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.2.3
